% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_builder.R
\name{build_report_data_json}
\alias{build_report_data_json}
\title{Build complete JSON data object for JavaScript}
\usage{
build_report_data_json(report_data)
}
\arguments{
\item{report_data}{List of pre-calculated report data from prepare_report_data()}
}
\value{
Character JSON string
}
\description{
Serializes pre-calculated report data to JSON format for JavaScript consumption.
This is a pure serialization function - all business logic should be done
in prepare_report_data() before calling this function.
}
