% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_num_participants}
\alias{calculate_num_participants}
\title{Calculate total number of participants}
\usage{
calculate_num_participants(metrics)
}
\arguments{
\item{metrics}{List of metric data frames from parse_delivery_metrics()}
}
\value{
Integer participant count
}
\description{
Sums all rows in person table (valid + invalid) to get participant count.
}
