% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsers.R
\name{create_empty_metric_df}
\alias{create_empty_metric_df}
\title{Create empty metric data frame with correct structure}
\usage{
create_empty_metric_df(parser_config)
}
\arguments{
\item{parser_config}{Parser configuration}
}
\value{
Empty data frame with correct columns
}
\description{
Create empty metric data frame with correct structure
}
