% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsers.R
\name{format_date_safe}
\alias{format_date_safe}
\title{Helper: Format date safely}
\usage{
format_date_safe(date_raw)
}
\arguments{
\item{date_raw}{Raw date string}
}
\value{
Formatted date string
}
\description{
Helper: Format date safely
}
