% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcs_authenticate}
\alias{gcs_authenticate}
\title{Authenticate to GCS using service account credentials}
\usage{
gcs_authenticate()
}
\value{
Logical TRUE if authentication succeeded, FALSE otherwise
}
\description{
Authenticates using Cloud Run service account at /secrets/service-account.json
or local default credentials.
}
