% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_order}
\alias{get_table_order}
\title{Get canonical display order for tables}
\usage{
get_table_order()
}
\value{
Character vector of table names in display order
}
\description{
Get canonical display order for tables
}
