% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type_concept_colors}
\alias{get_type_concept_colors}
\title{Get type concept color mapping}
\usage{
get_type_concept_colors()
}
\value{
Named character vector of hex colors
}
\description{
Get type concept color mapping
}
