% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_type_concept_group_order}
\alias{get_type_concept_group_order}
\title{Get canonical order for type concept groups}
\usage{
get_type_concept_group_order()
}
\value{
Character vector of group names in display order
}
\description{
Get canonical order for type concept groups
}
