% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_vocabulary_tables}
\alias{get_vocabulary_tables}
\title{Get vocabulary tables}
\usage{
get_vocabulary_tables()
}
\value{
Character vector of vocabulary table names
}
\description{
Get vocabulary tables
}
