% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_escape}
\alias{html_escape}
\title{Escape HTML special characters}
\usage{
html_escape(x)
}
\arguments{
\item{x}{Character vector to escape}
}
\value{
Character vector with &, <, > replaced by their HTML entities
}
\description{
Escape HTML special characters
}
