% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gcs_path}
\alias{is_gcs_path}
\title{Check if path is a GCS path}
\usage{
is_gcs_path(path)
}
\arguments{
\item{path}{Character path to check}
}
\value{
Logical TRUE if path starts with "gs://"
}
\description{
Check if path is a GCS path
}
