% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{is_harmonized_table}
\alias{is_harmonized_table}
\title{Check if a table participates in vocabulary harmonization}
\usage{
is_harmonized_table(table_name)
}
\arguments{
\item{table_name}{Character table name}
}
\value{
Logical TRUE if table is harmonized
}
\description{
Check if a table participates in vocabulary harmonization
}
