% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_renderer.R
\name{load_template}
\alias{load_template}
\title{Load a template file from inst/templates}
\usage{
load_template(template_name)
}
\arguments{
\item{template_name}{Template path relative to inst/templates (e.g., "sections/header")}
}
\value{
Character string containing template content
}
\description{
Load a template file from inst/templates
}
