% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_tables_by_order}
\alias{sort_tables_by_order}
\title{Sort table names according to canonical display order}
\usage{
sort_tables_by_order(tables)
}
\arguments{
\item{tables}{Character vector of table names to sort}
}
\value{
Character vector of sorted table names
}
\description{
Tables in .TABLE_ORDER will be sorted according to that order.
Tables not in .TABLE_ORDER will appear alphabetically after.
}
