% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate results file schemas}
\usage{
validate_schema(data, results_file_type)
}
\arguments{
\item{data}{Data frame to validate}

\item{results_file_type}{Character containing the type of file to validate}
}
\value{
Logical TRUE if valid (or stops with error)
}
\description{
Checks that all required columns are present in specified file.
}
