% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_renderer.R
\name{get_logo_data_uri}
\alias{get_logo_data_uri}
\title{Convenience wrapper to load a Connect brand logo as a data URI}
\usage{
get_logo_data_uri(name = c("white", "favicon"))
}
\arguments{
\item{name}{Logo identifier: "white" (for dark backgrounds) or "favicon"}
}
\value{
Character string containing a complete data: URI
}
\description{
Convenience wrapper to load a Connect brand logo as a data URI
}
